+++ app/src/main/kotlin/com/arishawke/asala/calendar/ui/year/YearDayCell.kt
@@ -29,6 +29,7 @@ import androidx.compose.ui.draw.clip
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.res.pluralStringResource
 import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.semantics.Role
 import androidx.compose.ui.semantics.contentDescription
 import androidx.compose.ui.semantics.semantics
 import androidx.compose.ui.semantics.stateDescription
@@ -72,7 +73,7 @@ internal fun YearDayCell(
     Box(
         modifier = Modifier
             .aspectRatio(1f)
-            .clickable(enabled = inMonth, onClick = onClick)
+            .clickable(enabled = inMonth, role = Role.Button, onClick = onClick)
             .then(
                 if (inMonth) {
                     Modifier.semantics(mergeDescendants = true) {